feat: abstract growthbook context into own file
